mv wasm cache access trait and impls into core
